pyproject.toml
[project]
name = "mcp-fivetran"
version = "0.1.0"
description = "MCP Server for FiveTran"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "fivetran-connector-sdk>=1.4.0",
    "mcp[cli]>=1.7.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[project.scripts]
mcp-fivetran = "mcp_fivetran.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]
